refactor: change folder structure and set path
