+++ public/examples/shoot/Common/HealthBar.js
