+++ public/examples/shoot/Common/StateMachine.js
