+++ public/examples/shoot/Monster/MonsterStore.js
